Add a basic map widget
This is a very preliminary stab at a map widget using the OpenLayers
libraries from the web.  The current approach is not suitable for
long-term use since it requires the Internet, but it is a first step
towards the full stand-alone solution.  In addition to that, more
flexibility in rendering, clipping to specific towns/zones/bats,
etc. need to be added.