Load zone/BAT mapping at server startup
We need the mapping between BATs and zones for multiple purposed in the
dashboard.  This commit provides this information by reading it at
server startup from a simple JSON file formatted as a list of lists
where each sublist has just two elements, a BAT number followed by a
zone number.  The mapping is stored in the application data structure
for easy access from handlers.