Significantly refactor DASummary handlers
After getting basic map display and information layout working in the
DASummary handlers it was clear that there had to be a more concise and
readable implementation.  This commit refactors much of the code to make
progress toward that goal.  The biggest changes are adding type safety
to the map widget and extracting the navbar from the widgets for the
primary contents of each page type.  With those changes handler could
also be streamlined.