Display IP address and shutdown info in title bar
We'll be operating the dashboard in locations where the hostname(s) and
IP address(es) of the server might not be fixed and well-known, so print
that information in the title bar of the site.  The haskell code does
not try to figure this out but just expects the environment to contain
already populated information in the WIFI_ADDRESS and ETH_ADDRESS
variables.

Shutdown of the dashboard server is handled through a window manager key
sequence on the console.  This commit also displays that key sequence so
it is easily discoverable.