Add placeholder zone statuses element to App
The primary purpose of the dashboard is to display summaries of the
status of various areas in the city, referred to as BATs and Zones.
This commit adds the type instance to the application that will provide
the underlying data parsed from damage assessment message files in a
TVar so that it can be updated over time.