Make BatZoneStatuses live update
This commit adds a System.FSNotify watcher to BatZoneStatuses that
ensures that a TVar holding a BatZoneStatuses record is updated whenever
the data changes; this is currently done by re-computing the
BatZoneStatuses records from scratch, but it may, in the future, be
possible to use some kind of incremental updating.  This commit also
makes the application start one of these watchers in a new thread,
pointed at the TVar containing the BatZoneStatuses record for the whole
server, which ensures that any request to a handler will result in the
appearance of updated values.