Reintroduce summary page background status loading
Somehow this disappeared from the working directory and never got
committed.  We realized that the status was not in fact updating when we
were testing the instructions for building and running a container for
the server and found the missing code in the temporary autogen
Javascript files.

With this change automatic updating should be working for the summary
page, apparently for the first time in the commit history of the
official repository.  Hopefully we didn't miss other changes that
improve on this.