Pin nixpkgs and pack-it-forms-msgfmt for release builds
nixpkgs is pinned to the tip of the 17.09 branch at the time of this
commit. The intention is to keep it on the most recent stable commit
at each release.

pack-it-forms-msgfmt is pinned to the tip of master at the time of
this commit. It is expected that it will co-evolve with the dashboard
and is likely to be updated fairly frequently.